Switched to pythonic dbt assets
This was a fantastic change. Allowed for easy pooling of dbt resources,
less yaml magic and hacks, and the performance on backfills was awesome.
When performing a combined backfill, it looks like it divides up the dbt work
into dynamic partition ranges as it fills the dlt sources up.

Overall this was a much easier time than the components approach.