Update realism of call center simulation
This change makes the default setting for number of customers to be a function of the
number of agents and the number of calls per agent per day. This allows for some frequent
callers by chance and adds more realistic scenarios where a customer doesn't repeatedly call
often.

Made a change to prevent overlap of customers as well, so that a customer cannot be
simultaneously have calls with another agent. Also cleaned up the manage.py cli a bit.