Staging dbt updates
Adding a warehouse_updated_ts field to tables modeled in dbt.
Also removed the lookback window on stg_surveys because all the
received surveys on a date are already landed in the directory.
When we join survey metrics to a downstream model based on call/sent
date, then we will need to implement the lookback window.